Refactor TopProjects component to shared
