+++ packages/lib-user/src/components/shared/ContentBox/ContentBox.stories.js
@@ -100,7 +100,7 @@ export const TopProjectsFullWithFewerProjects = {
   )
 }
 
-export const TopProjectsSplit = {
+export const TopProjectsGrid = {
   render: () => (
     <ContentBox
       linkLabel='See more'
