+++ packages/lib-user/src/components/shared/index.js
@@ -10,3 +10,4 @@ export { default as ProfileHeader } from './ProfileHeader'
 export { default as Select } from './Select'
 export { default as Tabs } from './Tabs'
 export { default as TitledStat } from './TitledStat'
+export { default as TopProjects } from './TopProjects'
